+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - General:
   - Check file creation time and fingerprint when before reloading
 - Bug fixes:
-  - Fixed hang when opening multiple files
+  - Fixed opening multiple files
     
 ## v0.1.3
 - General: